+++ demo/client/src/date-range-picker.js
@@ -1,0 +1,13 @@
+import {initDatepicker, selectRef} from 'metronic-extension';
+import highlight from '~/shared/highlight';
+
+highlight();
+const ref = selectRef();
+initDatepicker(ref.dateRangePicker, {
+  minDate: moment().format('YYYY/M/D'),
+  locale: 'en',
+  language: {
+    applyLabel: 'OK',
+    cancelLabel: 'Cancel',
+  }
+});