+++ docs/update-docs.js
@@ -1,0 +1,3 @@
+/**
+ * This is a script that re-creates the docs based on the demo.
+ */